pyproject.toml
# "For legacy reasons, if a build system is not defined, then
# setuptools.build_meta:__legacy__ is used to build the package."
# https://docs.astral.sh/uv/concepts/projects/config/#build-systems

# We use Hatchling for reproducible builds
# The uv build backend, uv_build, still has reproducibility issues
# E.g. https://github.com/astral-sh/uv/issues/13139
[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[project]
name            = "apache-trusted-releases"
version         = "0.20251022.1835"
description     = "ATR CLI and Python API"
readme          = "README.md"
requires-python = ">=3.13"
license         = "Apache-2.0"
authors         = [{name = "ASF Tooling", email = "dev@tooling.apache.org"}]
classifiers     = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.13",
  "Operating System :: OS Independent",
]

dependencies = [
  "aiohttp",
  "cyclopts",
  "filelock",
  "pgpy>=0.6.0",
  "platformdirs",
  "pydantic",
  "pyjwt",
  "sqlalchemy",
  "sqlmodel",
  "standard-imghdr>=3.13.0",
  "strictyaml",
]

[dependency-groups]
dev = [
  "aioresponses",
  "pre-commit",
  "pytest-asyncio",
  "pytest-console-scripts",
  "pytest",
]

[project.scripts]
atr = "atrclient.client:main"

[tool.hatch.build.targets.wheel]
packages = ["src/atrclient"]

[tool.ruff]
line-length = 120
extend-exclude = [
    "node_modules",
]

[tool.ruff.lint]
ignore = []
select = [
  "C90",
  "E",
  "F",
  "I",   # isort
  "N",   # pep8-naming
  "RUF", # ruff-checks
  "TC",  # flake8-type-checking
  "TID", # flake8-tidy-imports
  "UP",  # pyupgrade
  "W"
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:imghdr was removed in Python 3.13:DeprecationWarning:pgpy.constants",
]

[tool.uv]
exclude-newer = "2025-10-22T18:35:00Z"
